Create a persistent connection that can be used to send multiple push notifications.
This cleans up the implementation in client.go and supersedes it. However, for backward-compatibility, client.go will remain for a while. Apple treats a lot of connects/disconnects as a potential DoS attack; since nobody knows the threshold, you should use one connection to send multiple notifications.

The connection also allows the timeout to be configured by the user; it defaults to the TIMEOUT_SECONDS constant but provides a SetTimeoutSeconds method.